#2568 Behat context
